added additional interfaces matching implementations
